requirements.txt
altgraph==0.17.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
Deprecated==1.2.18
Flask==3.1.0
Flask-Limiter==3.10.1
Flask-Session==0.8.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
limits==4.0.1
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.19.0
numpy==2.2.3
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pycparser==2.22
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
threadpoolctl==3.5.0
tk==0.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==2.1.4
